feat(core/binding): fix some issues
